2319: Merge `ApplicationStatusBarItem.tsx` into `ApplicationStatusBar.tsx`
